FoodTourVinhKhanh/Backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
passlib[bcrypt]
bcrypt==4.0.1
python-jose
python-multipart
email-validator
python-dotenv
redis==7.3.0
requests==2.33.0